move metrics macro to fuel-indexer-macros
